fix memory leak on job execution from baseservice (#240)
The `BaseService` type was originally only used for long-running
internal services. At some point, we refactored the job execution code
to also make use of this type. That exposed a memory leak: every time a
new `BaseService` is `Init`'d, a new random source is allocated. That
meant for every single job.

This is of course totally unnecessary, and to avoid it we can move the
rand source to the `Archetype`, which is only allocated once per client.

Partially fixes #239.