Periodic job enqueuer: Keep timer paused around inserts to fix intermittency
Here, aim to fix the intermittent test problem described in #215, in
which after inserting a periodic batch where periodic 500 ms jobs and
periodic 1500 ms jobs should've lined up, we come up with fewer 500 ms
jobs than expected, suggesting that only a 1500 ms job was inserted
during the run.

    --- FAIL: TestPeriodicJobEnqueuer (0.00s)
        --- FAIL: TestPeriodicJobEnqueuer/EnqueuesPeriodicJobs (1.57s)
            logger.go:225: time=2024-02-19T07:53:42.969Z level=INFO msg="PeriodicJobEnqueuer: Run loop started"
            periodic_job_enqueuer_test.go:112:
                    Error Trace:	/home/runner/work/river/river/internal/maintenance/periodic_job_enqueuer_test.go:82
                                                            /home/runner/work/river/river/internal/maintenance/periodic_job_enqueuer_test.go:112
                    Error:      	"[%!s(*dbsqlc.RiverJob=&{45 [1[23](https://github.com/riverqueue/river/actions/runs/7956225078/job/21716534964?pr=214#step:8:24) 125] 0 <nil> [] {482531000 63843926023 0x107e100} [] <nil> periodic_job_500ms 25 [123 125] 1 default available {482709000 63843926023 0x107e100} []}) %!s(*dbsqlc.RiverJob=&{46 [123 125] 0 <nil> [] {983499000 63843926023 0x107e100} [] <nil> periodic_job_500ms 25 [123 125] 1 default available {983650000 63843926023 0x107e100} []})]" should have 3 item(s), but has 2
                    Test:       	TestPeriodicJobEnqueuer/EnqueuesPeriodicJobs
                    Messages:   	Expected to find exactly 3 job(s) of kind: periodic_job_500ms, but found 2
            logger.go:225: time=20[24](https://github.com/riverqueue/river/actions/runs/7956225078/job/21716534964?pr=214#step:8:25)-02-19T07:53:44.472Z level=INFO msg="PeriodicJobEnqueuer: Run loop stopped"
    FAIL
    FAIL	github.com/riverqueue/river/internal/maintenance	1.761s

As with many of these intermittent problems, this one is quite difficult
to fully confirm a fix on because I'm finding it basically impossible to
reproduce locally, regardless of the number of test iterations I run.
However, after reauditing the code, I think I know what the problem was.

Previously, the enqueuer's run loop worked by initializing a timer with
a zero reset interval once outside the loop, then immediately resetting
the timer to the actual target run interval at the top of every loop
iteration.

A slightly non-obvious side effect of this implementation is that the
timer would run _while_ the loop did its work inserting periodic jobs.
Usually inserting periodic jobs (especially in this test) is very fast
and took only a millisecond or two, so the effect of running the timer
simultaneously was negligible, but in extremely degenerate environments
like GitHub Actions, there may be long delays in any goroutine, and it's
possible that there was a significant desynchronization between when the
timer was scheduled to fire next, and when inserted jobs were inserted
at. I believe that's what our problem was -- it's an unlikely race, but
given a runtime with frequent pauses, one that could conceivably happen.

Here, modify the implementation so that the initial timer initialization
is set with an actual TTL until the next periodic job (instead of zero),
and then move the timer reset until after periodic jobs have been
inserted in the loop. This has the effect of keeping the timer entirely
paused while periodic jobs are being inserted, and (at least I think)
reducing the probability of a timer/insert desynchronization to zero.

It also makes the code a bit cleaner and easier to read IMO, so a side
bonus.

Fixes #215.