Make notifier operations more deterministic + fix intermittency
I have a hypothesis that a lot of our test intermittency in high level
clien tests stem from deep down in the notifier. It works the vast
majority of the time, but looking into its implementation, it's
currently written to be non-deterministic. When using `Listen` to add a
new subscription, the notifier only attempts to do so asynchronously by
putting it in a queue for eventually processing. Callers don't know
whether adding a subscription succeeded, and more importantly (as far as
our intermittency problems are concerned), don't know _when_ it may have
succeeded.

Here, we make quite a few cosmetic changes, but the most important one
is to the API and implementation of `Listen`:

    sub, err := notifier.Listen(...)

Notably, it:

1. Now returns an error. If there was a problem listening, the caller is
   told about it.

2. Operations synchronously. `Listen` waits until a subscription could
   be confirmed, and then returns it, with the returned subscription
   known to be working immediately, as opposed to an indeterminate time
   in the future.

The synchronous works by having `Listen` schedule an "interrupt", so if
the service is it its main loop blocking on `WaitForNotification`, the
wait is cancelled, the loop processes any pending listens and unlistens
in interrupts, then it reenters `WaitForNotification` once again.

We add a much more comprehensive test suite to help verify various
behavioral edges, and suss out problems via use of `-race`. I believe I
found quite a few other smaller bugs in the notifier while doing this,
but admittedly things got a bit muddied because I surely introduced many
of those bugs during my main refactor.

A few other miscellaneous changes:

* The notifier becomes a `startstop`. Along with benefits around
  preventing double starts and many concurrency problems, it also
  changes to be able to error quickly in case of an obvious immediate
  problem like an unusable connection pool.

* We implement a true exponential backoff wait in case of a bad error in
  the main loop.

* Try to improve logging so that in case of future test failures, we
  should have a fairly good trace to see what happened when leading up
  to a failure.

To confirm/disconfirm resolution of the intermittency problems, I've
been running the infamous:

    $ go test . -run Test_Client_InsertTriggersImmediateWork -race -count 5000

It's had about 10 successes now. I did run into one intermittent failure
on one run, but it was a brand new one in the client that's not the
usual we see in CI, and I think unrelated to any changes made here.