Generate default client IDs using host and client start time (#255)
Here, move away from ULIDs for generating client IDs, and over to a
custom ID generator that uses a combination of host and client start
time. This has two benefits:

* Host/start time are more descriptive for humans. It should give an
  operator a rough idea as to where the name of a client originated
  from, and if it's a client that's likely long dead (if the timestamp
  is very old).

* Lets us drop our ULID package dependency, which we don't use for
  anything else.

Generated IDs look like this for example:

    miniml2_local_2024_03_07T04_39_12

I've tried to arrange them so that on most systems you'll be able to
double click on an ID to select the whole thing, which might be helpful
while copying information or debugging a production problem. Characters
like colons (`:`) are less friendly for double-click-to-select, so I've
avoided them. Hyphens and dots aren't universally friendly for double
click either (depends on program and configuration), so I've avoided
those as well.