Also allow a deallocation error from pgx during `TestTx` cleanup
This is a new error that's appeared as of pgx 5.5.5 [1] wherein the
"conn closed" error we were handling can now be wrapped by a "failed to
deallocate cached statement(s)" error, neither of which has a fixed
error variable to use with `errors.Is`.

The build initiated by Dependabot failed due to this problem, so here we
push a change that allows it to succeed.

[1] https://github.com/riverqueue/river/pull/265