Fix intermittent notifier stress test (#261)
Here, fix an intermittent stress test in the notifier package. This is
one that I added last minute in #258 in response to code review, so I
missed the intermittency because it didn't fail on any of the last
pushes that I made, but I noticed it failing today [1].

Luckily, it's quite a simple where we just have to make sure that the
test waits for one of the test goroutines to return before quitting. Ran
at `-race -count 1000` locally and no longer seeing any intermittency.

[1] https://github.com/riverqueue/river/actions/runs/8223455278/job/22486174354