Make more client services into `startstop.Service` (#262)
Here, we aim to move in a direction where the client can treat most of
its subservices generically by starting/stopping the notifier as a
generic service, along with converting three other smaller services over
to use the `startstop.Service` interface:

* Client monitor. (This ends up cleaning up its internal code
  significantly, and has a major improvement on its test coverage and
  robustness.)

* The client's statistics logging loop.

* The client's leadership change handling loop.

Both the latter two are very small functions, so I implement the new
function `startstop.StartStopFunc` which similar to `river.WorkFunc`,
provides an easy way to get a small service using only a function, which
keeps the code quite clean.

It's going to take a few more changes to get our full simplification
change, but you can already see the beginnings of it as many services
can be started and stopped in ~one line:

    func (c *Client[TTx]) signalStopComplete(ctx context.Context) {
        // Wait for producers and elector to exit:
        c.wg.Wait()

        // Stop all mainline services where stop order isn't important.
        startstop.StopAllParallel(c.services)

By the end of it, both the elector and producers will also become
services, and we'll be able to fully eliminate the wait group, along
with the stop channels and the roundabout stop functions like
`signalStopComplete`. The idea is that the client itself will also
become a start/stop service, and itself become protected against being
started multiple times, or potential races where `Start`/`Stop` may be
called by multiple goroutines.