Make producer start/stop service + poll-only mode + expanded tests (#264)
Follow up changes like #253, #262, and #263 to make the producer a
start/stop service, giving it a more predictable way to invoke start and
stop, making it safer to run and cleaning up caller code where it's used
in the client and test cases. With all these changes taken together
we'll have every service in the client using the same unified service
interface, which will clean up code and let us write some neat utilities
to operate across all of them.

Aside from that, we clean up the producer in ways to bring it more
inline with other code, like making logging uniform and having the
constructor return only a `*producer` instead of `(*producer, error)`
that needs to be checked despite an error always being indicative of a
bug in this context.

We expand the test suite, adding tests like (1) verifying that workers
are really stopped when `workCtx` is cancelled, (2) verifying that the
max worker slots work as expected and that the producer limits its
fetches, and (3) start/stop stress.

Like with #263, we give the producer a poll only mode, which also gets
the full test barrage using a shared test transaction instead of full
database pool. Also like #263, this poll only mode is still prospective
and not yet put to full use (although it will be soon).