Fix build status badge in README (#267)
Rendering of the build status badge in the README on GitHub is currently
broken because of the change I made in #228 that changed the extension
of the CI file from `.yml` to `.yaml`. I didn't realize the filename was
in the badge's source path, but it is, and trying to locate `ci.yml` no
longer finds anything. Here, a very small change to fix the problem.