Fix intermittent test failure on periodic job enqueuer by raising "ready margin"
This one's aimed at fixing the intermittent test described in #215. By
reading some additional logging probes in #271, I've been able to see
that the failures occur when the two jobs in the test case, 500ms and
1500ms, have their ready times diverge ever so slightly across run
iterations, which given enough test runs, can diverge enough so that
both jobs aren't ready on the last check where they're supposed to sync
up.

There's a few different potential ways to solve this one, most involving
rewriting the test, but here I'm proposing that we solve it by
increasing the "now margin" in the enqueuer from 10ms to 100ms. This is
a small margin applied on each run loop that look for jobs that are not
quite ready, but so close to ready that we enqueue them anyway. The
choice of 10ms originally was somewhat arbitrary, and 100ms isn't a
substantially larger number and still makes sense, so I think this is a
reasonable resolution.

I verified the fix works by pushing it up in the repro bootstrap in #271
that runs with `-race` and high iteration count in the slower GitHub CI
environment. It moves the CI matrix from failing on every job on every
run to succeeding on every job on every run.

Fixes #215.