Make more River logging debug level
Here, in an attempt to reduce the level of noise of River logging
somewhat, move more logging that's not as useful from into to debug
level. Most of the changed log statements are the start/stop message of
each subservice (so the River client will still emit start/stop at info
level, but each subservice moves theirs to debug).

Also remove a few log statements that are quite duplicative of an almost
identical one happening elsehwere (e.g. "client started" in benchmark,
which is almost the same as "client started" in the client), and a few
that I'm pretty sure I'd left in by accident while debugging.

The CLI now takes a `--debug` parameter, and `riverinternaltest` now
provides `BaseServiceArchetypeDebug` to easily swap into a failing test,
so access to the additional logging is fairly easy to get at in case
it's needed.