Tee up version `v0.1.0` including new CLI release
Tees up version `v0.1.0`, containing a number of changes like `river
bench`, `river migrate-get`, and a new, faster, completer.

We could pool a few more changes before release, but I kind of want to
get this out the door so I don't forget to the cut a CLI release and
remove the `replace` directives in its `go.mod`.