Fix hot loop in reindexer service + no more disabling sleep in base services
Fix a bug in the reindexer which was causing it to enter a hot loop on
shutdown. It wasn't doing any work since reindexing is currently
disabled, but it was emitting an inordinate amount of logging output.

The problem stemmed from trying to use a `context.Deadline` as a timer:

    for {
        timerCancel := context.WithDeadline(ctx, nextRunAt)

        select {
        case <-ctx.Done():
            timerCancel()
            return
        case <-timerCtx.Done():
        }

        ...
    }

The bug is easy to miss on the first read, but involves context
inheritance. A service's context is cancelled when its being stopped,
and in this case `timerCtx` inherits from `ctx`, so on stop, _both_
`ctx` and `timerCtx` are cancelled, making the `select` statement's
precedence indeterminate, and causing the loop to fall through and
continue an arbitrary number of times. (I would've expected the branch
taken to be a 50/50, but there's some other implementation detail at
work here because I was seeing `timerCtx` preferred in a statistically
improbable way on most runs, and it'd often fire 4-5 times before
`ctx.Done()` was found a test case would end.)

I replaced the use of a context with a `CancellableSleep` with a context
check following it.

This then let to a bit of a yakshake because the cancellable sleep was
being respect _either_, which turned out to be because all base service
sleep is disabled in the queue maintainers in order to avoid their
staggered sleep that occurs on start up.

I changed my mind a while ago that being able to broadly disable sleep
like this is actually not a very good idea, but hadn't gotten around to
refactoring it out yet, so I finally did so here. The staggered sleep is
moved to a new base function with a disable flag specifically for this
use rather than all sleeps everywhere:

    func (s *queueMaintainerServiceBase) StaggerStart(ctx context.Context) {
        if s.staggerStartupDisabled {
            return
        }

        const (
            jitterMin = 0 * time.Second
            jitterMax = 1 * time.Second
        )

        s.CancellableSleepRandomBetween(ctx, jitterMin, jitterMax)
    }

The client then may disable start on all the queue maintainer services
as it initializes them based on a non-exported flag from its `Config`.
Unlike disabling sleep, disabling start up stagger is _always_
appropriate in client tests, and won't lead to unintended side effects
like accidentally disabling sleep on exponential backoffs.

Over all, despite a little more work for services that want to disable
their sleep for tests, I think this produces a far more obvious model
that's going to lead to fewer test bugs.