Use timer for reindexer instead of cancellable sleep
Quick follow up to #280 in which we rebuild the reindexer somewhat to
look more like the periodic job enqueuer. The benefits in readability of
each is debatable, but the periodic job enqueuer's timer model has the
minor advantage of reusing the same timer over and over again, and also
that it makes the difference maintenance services look more like each
other. It also allowed the opportunity to simplify the code somewhat and
removed one of the time variables which is no longer needed.

Also, change a line in the periodic job enqueuer so we use a `return`
statement instead of a break with a named loop.