Periodic job enqueuer: Use original next run time when calculating next schedule
This one came up while I was looking into #283. Although I don't think
it'll fully resolve what's being reported there, this implementation
feels a little more correct compared to what we had originally.

Currently, when scheduling the next run of a job, we use the current
time. However, the current time may be somewhat staggered from when the
job was supposed to run at because we build in a 100 ms "ready margin"
when looking for jobs to run. Although a minor problem, this makes it
conceivably possible for a periodic job to drift 100 ms at a time until
it's deviated from its theoretically ideal run schedule by a non-trivial
margin (well, debatably, your definition of trivial depends a lot on
schedule tolerance).

Here, make a minor change that uses the job's original run target time
when calculating the next one. This will have been within 100 ms of the
current time anyway, but is a little more accurate for what was intended.

I started writing a test case for this, but anything I put in is so
intrusive that it feels like the noise added to code is bad enough that
it outweighs any additional marginal benefit from the test, so I just
opted not to have a test. The existing ones don't differentiate a
schedule based on current time versus original run time, but they do vet
that the change still works.